JIRA: RW-845 Remove dead function.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@92373 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/readers/gtf_location_merger.cpp
@@ -44,7 +44,7 @@ CGtfLocationRecord::CGtfLocationRecord(
     CGff3ReadRecord::SeqIdResolver seqIdResolve)
 //  ----------------------------------------------------------------------------
 {
-    static map<string, size_t> typeToPart {
+    static map<string, int> typeToPart {
         {"cds",         1},
         {"start_codon", 0},
         {"stop_codon",  2},
